Внесение изменений рекомендованных ревьюером:
SearchActivity: Вынесена логика сохранения истории в отдельный класс HistoryManager
